+++ cluster-controller/src/main/java/io/strimzi/controller/cluster/operations/resource/UpdateS2IOperations.java
@@ -1,8 +1,6 @@
-package io.strimzi.controller.cluster.operations.openshift;
+package io.strimzi.controller.cluster.operations.resource;
 
 import io.fabric8.openshift.client.OpenShiftClient;
-import io.strimzi.controller.cluster.operations.resource.BuildConfigOperations;
-import io.strimzi.controller.cluster.operations.resource.ImageStreamOperations;
 import io.strimzi.controller.cluster.resources.Source2Image;
 import io.vertx.core.Future;
 import io.vertx.core.Vertx;
@@ -16,16 +14,16 @@ import java.util.List;
 /**
  * Updates all Source2Image resources
  */
-public class UpdateS2IOperation extends S2IOperation {
+public class UpdateS2IOperations extends S2IOperations {
 
-    private static final Logger log = LoggerFactory.getLogger(UpdateS2IOperation.class.getName());
+    private static final Logger log = LoggerFactory.getLogger(UpdateS2IOperations.class.getName());
     private final BuildConfigOperations buildConfigOperations;
     private final ImageStreamOperations buildImageStreamOperations;
 
     /**
      * Constructor
      */
-    public UpdateS2IOperation(Vertx vertx, OpenShiftClient client) {
+    public UpdateS2IOperations(Vertx vertx, OpenShiftClient client) {
         super(vertx, "update");
         buildConfigOperations = new BuildConfigOperations(vertx, client);
         buildImageStreamOperations = new ImageStreamOperations(vertx, client);
